+++ instrumentation/cmplog-switches-pass.cc
@@ -18,6 +18,7 @@
 #include <stdio.h>
 #include <stdlib.h>
 #include <unistd.h>
+#include <stdint.h>
 
 #include <iostream>
 #include <list>
@@ -143,6 +144,16 @@ bool CmplogSwitches::hookInstrs(Module &M) {
   FunctionCallee c8 = M.getOrInsertFunction("__cmplog_ins_hook8", VoidTy,
                                             Int64Ty, Int64Ty, Int8Ty);
   FunctionCallee cmplogHookIns8 = c8;
+
+#if INTPTR_MAX != INT32_MAX
+  IntegerType   *Int128Ty = IntegerType::getInt128Ty(C);
+  FunctionCallee c16 = M.getOrInsertFunction("__cmplog_ins_hook16", VoidTy,
+                                             Int128Ty, Int128Ty, Int8Ty);
+  FunctionCallee cmplogHookIns16 = c16;
+  FunctionCallee cN = M.getOrInsertFunction("__cmplog_ins_hookN", VoidTy,
+                                            Int128Ty, Int128Ty, Int8Ty, Int8Ty);
+  FunctionCallee cmplogHookInsN = cN;
+#endif
 
   GlobalVariable *AFLCmplogPtr = M.getNamedGlobal("__afl_cmp_map");
 
@@ -304,7 +315,7 @@ bool CmplogSwitches::hookInstrs(Module &M) {
                 IRB.CreateCall(cmplogHookIns8, args);
                 break;
               case 128:
-#ifdef WORD_SIZE_64
+#if INTPTR_MAX != INT32_MAX
                 if (max_size == 128) {
 
                   IRB.CreateCall(cmplogHookIns16, args);