+++ test/test-pre.sh
@@ -95,6 +95,7 @@ unset AFL_PYTHON_MODULE
 unset AFL_PRELOAD
 unset LD_PRELOAD
 unset SKIP
+unset AFL_CMPLOG_ONLY_NEW
 
 rm -rf in in2 out